CustomArea uses GB::Category instead of ES data
