+++ app/models/gobierto_budgets/custom_area.rb
@@ -1,6 +1,10 @@
 module GobiertoBudgets
   class CustomArea < BudgetArea
 
+    def self.any_items?(options = {})
+      GobiertoBudgets::Category.where(site: options[:site], kind: options[:kind]).any?
+    end
+
     def self.area_name
       'custom'
     end
